Improve robust/correctness of reproducibility test
Previously eval_funcs could have passed by simply returning the evaluators passed to them; this would always give a false negative, as the second run of the evaluation function can mutate the evaluators; the test was too sensitive to the implementation.
This change explicitly constructs separate evaluators to pass to the function, and observes the results of the evaluators passed to and returned from the eval_func, ensuring that both runs of the eval_func yield the same results regardless of whether or not it constructs new evaluators.