Add sanity test of bootstrapping (STOCHASTIC)
Note; this test is 'stochastic' in nature, but will be deterministic as long as random seeds are used judiciously. There will be a ~5% false failure rate if something is changed downstream.

A reasonable fix is to run the test multiple times on different seeds to reduce the chance of a false failure